requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==3.2.12
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
wincertstore==0.2
django-extensions==3.1.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
django-environ==0.8.1